refactor: move TripAdvisorResponse model to separate class
